LayoutTests/fast/viewport/viewport-133.html

<head>
    <title>Check that device dimensions are not adapted due to the higher DPI.</title>
    <meta name="viewport" content="width=device-width, target-densitydpi=high-dpi">
    <script src="../../fast/js/resources/js-test-pre.js"></script>
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(240, 480, 564, 480, 534);
                shouldBeCloseTo("window.devicePixelRatio", 0.7083333134651184, 0.001, false);
            }
        }
    </script>
</head>
<body onload="test();">
