LayoutTests/fast/viewport/viewport-133-expected.txt

viewport size 480x534 scale 1.000000 with limits [1.000000, 5.000000] and userScalable -1.000000
PASS window.devicePixelRatio is within 0.001 of 0.7083333134651184
